requirements.txt
aiosmtplib==3.0.2
alabaster==1.0.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.0
cryptography==45.0.3
Deprecated==1.2.18
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-mail==1.5.0
greenlet==3.2.2
h11==0.16.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
libgravatar==1.0.4
limits==5.2.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
redis==6.2.0
requests==2.32.4
roman-numerals-py==3.1.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
starlette==0.40.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wrapt==1.17.2
